+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.14.1
+### Removed
+- Dump slow protobuf serialization.
+
+### Added
+- Add custom serialization.
+
+### Changed
+- Bump `haskoin-core` nad `haskoin-node`.
+
 ## 0.13.1
 ### Changed
 - Bump `haskoin-node` in `stack.yaml`.